Restructured Timer ISR, added tasks and host mode, fixed bugs.
Events.c and ChipInit.c: Moved UI update out of task block, disabled
RotEnc interrupt unless at menuLevel 3.
utilityFuncts.c: Fixed length bug in fillBuffer function.
UserInterface.c: Added more uiChange statements.
Added Tasks.c.

+++ JsonFuncts.h
@@ -14,7 +14,7 @@
 #include <ctype.h>
 #include <avr/pgmspace.h>
 #include "macros.h"
-int jsonBuffer2UiMenu(void);
+uint8_t jsonBuffer2UiMenu(void);
